New Ensemble stuff! Now instead of shinanigans that don't make sense to me anymore (old Ensemble.py) we have EnsembleMask.py.
This new ensemble method uses a mask for each image using the detection, model, and confidence in a semi-organized fassion to weight the mask regions.
This scoring/weighting requires some parameters that are not tuned yet
Then we find contours in the mask and convert them to bounding boxes using the mean confidene of all the pixels as the new confidence.

Other updates:
- fixed nms function